[sk] Refactored to avoid .lark dependency
