+++ mage_ai/data_cleaner/transformer_actions/row.py
@@ -1,7 +1,7 @@
 from mage_ai.data_cleaner.column_types.constants import NUMBER_TYPES
 from mage_ai.data_cleaner.transformer_actions.action_code import query_with_action_code
 from mage_ai.data_cleaner.transformer_actions.custom_action import execute_custom_action
-from mage_ai.data_cleaner.transformer_actions.query import QueryGenerator
+from mage_ai.data_cleaner.transformer_actions.query.query import QueryGenerator
 import pandas as pd
